requirements.txt
hachoir==3.1.2
Pillow>=9.0.0
psutil==5.8.0
pymongo==3.12.0
Pyrogram==2.0.26
pyromod==1.5
telegram
TgCrypto
dnspython
ffmpeg-python
